refactor: move the functions from controllers to index.ts
